Move get_slurm_config module to v2 scope
